pyproject.toml
[project]
name = "telegram-diary"
version = "0.1.0"
description = "Async Telegram diary bot with separate aiohttp web service and PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.16.0",
    "aiohttp>=3.12.0",
    "sqlalchemy>=2.0.37",
    "asyncpg>=0.30.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "psycopg2>=2.9.11",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
]
